Move buildImaTagUrl to /targeting/youtube-ima
