+++ src/ad-targeting-youtube.ts
@@ -1,21 +1,16 @@
-import type { Participations } from '@guardian/ab-core';
 import type { ConsentState } from '@guardian/consent-management-platform/dist/types';
 import { getCookie } from '@guardian/libs';
 import { canUseDom } from './lib/can-use-dom';
 import { constructQuery } from './lib/construct-query';
 import { getPermutivePFPSegments } from './permutive';
-import { filterEmptyValues } from './targeting/build-page-targeting';
-import { experimentsTargeting } from './targeting/session';
 import type {
 	AdsConfig,
 	AdsConfigBasic,
 	AdsConfigCCPAorAus,
 	AdsConfigDisabled,
 	AdsConfigTCFV2,
-	MaybeArray,
+	CustomParams,
 } from './types';
-
-type CustomParams = Record<string, MaybeArray<string | number | boolean>>;
 
 const disabledAds: AdsConfigDisabled = { disableAds: true };
 
@@ -92,78 +87,4 @@ const buildAdsConfigWithConsent = (
 	return buildAdsConfig(consentState, adUnit, customParamsToMerge);
 };
 
-/**
- * @param  {Record<string, MaybeArray<string|number|boolean>>
- * do this https://support.google.com/admanager/answer/1080597
- */
-const encodeVastTagKeyValues = (
-	query: Record<string, MaybeArray<string | number | boolean>>,
-): string => {
-	const unencodedUrl = Object.entries(query)
-		.map(([key, value]) => {
-			let queryValue: string;
-			if (Array.isArray(value)) {
-				queryValue = value.join(',');
-			} else if (typeof value == 'boolean' || typeof value == 'number') {
-				queryValue = value.toString();
-			} else {
-				queryValue = value;
-			}
-			return `${key}=${queryValue}`;
-		})
-		.join('&');
-	return unencodedUrl
-		.replace(/=/g, '%3D')
-		.replace(/&/g, '%26')
-		.replace(/,/g, '%2C');
-};
-
-const buildImaAdTagUrl = (
-	adUnit: string,
-	customParams: CustomParams,
-	clientSideParticipations: Participations,
-): string => {
-	console.log(clientSideParticipations);
-	const customParameters = {
-		ab: experimentsTargeting({
-			clientSideParticipations,
-			serverSideParticipations: window.guardian.config.tests ?? {},
-		}),
-		...customParams,
-		...buildCustomParamsFromCookies(),
-	};
-	console.log('raw cust params', customParameters);
-
-	const queryParams = {
-		iu: adUnit,
-		// iu: '/59666047/theguardian.com',
-		description_url: '[placeholder]', // do we need this?
-		tfcd: '0',
-		npa: '0',
-		sz: '480x360|480x361|400x300',
-		gdfp_req: '1',
-		output: 'vast',
-		unviewed_position_start: '1',
-		env: 'vp',
-		impl: 's',
-		correlator: '', // do we need this?
-		vad_type: 'linear',
-		vpos: 'preroll',
-		cust_params: encodeVastTagKeyValues(
-			filterEmptyValues(customParameters),
-		),
-	};
-
-	console.log(queryParams);
-
-	const queryParamsArray = [];
-	for (const [k, v] of Object.entries(queryParams)) {
-		queryParamsArray.push(`${k}=${v}`);
-	}
-	return (
-		'https://pubads.g.doubleclick.net/gampad/live/ads?' +
-		queryParamsArray.join('&')
-	);
-};
-
-export { buildAdsConfigWithConsent, buildImaAdTagUrl, disabledAds };
+export { buildAdsConfigWithConsent, disabledAds };
