+++ src/index.ts
@@ -38,11 +38,7 @@ export {
 export { initTrackScrollDepth } from './track-scroll-depth';
 export { initTrackLabsContainer } from './track-labs-container';
 export { initTrackGpcSignal } from './track-gpc-signal';
-export {
-	buildAdsConfigWithConsent,
-	buildImaAdTagUrl,
-	disabledAds,
-} from './ad-targeting-youtube';
+export { buildAdsConfigWithConsent, disabledAds } from './ad-targeting-youtube';
 export { createAdSlot, concatSizeMappings } from './create-ad-slot';
 export type {
 	AdsConfig,
@@ -78,3 +74,4 @@ export { a9Apstag } from './__vendor/a9-apstag';
 export { ipsosMoriStub } from './__vendor/ipsos-mori';
 export { launchpad } from './__vendor/launchpad';
 export { pubmatic } from './__vendor/pubmatic';
+export { buildImaAdTagUrl } from './targeting/youtube-ima';
